+++ tests/TeeForge.Experimental.Storage.Tests/GlobalUsings.cs
@@ -1,6 +1,6 @@
 global using TeeForge.Hashing;
 global using TeeForge.Mirroring;
-global using TeeForge.Pipelines;
+global using TeeForge.Broadcasting;
 global using TeeForge.RandomAccess;
 global using TeeForge.Experimental.Storage.Sparse;
 global using TeeForge.ErasureCoding.Internal;