+++ tests/TeeForge.Tests/GlobalUsings.cs
@@ -1,4 +1,4 @@
 global using TeeForge.Hashing;
 global using TeeForge.Mirroring;
-global using TeeForge.Pipelines;
+global using TeeForge.Broadcasting;
 global using TeeForge.RandomAccess;